refactor(hostd): move the supervisor self-measurement into the library
The measurement shipped in a binary that no CI lane compiles. mvm-hostd's
default feature set is empty and the bin is required-features = ["libkrun-sys"];
the one job that turns that feature on scopes to -p libkrun-sys and
-p mvm-runtime, and its path filter does not even list crates/mvm-hostd. The
release lane only builds the bin. So the first real CPU and memory reading in
this feature arrived with three tests that ran nowhere -- a witness that exists
while nothing executes it, which is the failure this repo keeps rediscovering.

Nothing about the measurement is libkrun-specific: a path, two mvm-vmm probes,
one file write. It moves to supervisor/self_usage.rs, beside wall_clock.rs,
which is the existing precedent for a feature-independent module that exists to
be shared by the per-VM supervisor binaries -- and is where the HVF supervisor,
the other in-process VMM tier, will reach for it.

The binary keeps what genuinely needs its context: the OnceLock state dir, the
extern "C" trampoline, and the atexit registration.

A third test comes with the move. The `let _ =` swallowing a failed sidecar
write is a load-bearing promise that a workload which already ran never loses
its process over evidence, and nothing checked it.

cargo nextest run -p mvm-hostd goes from 1920 tests to 1923: three that ran in
no lane at all now run in the ordinary one.
