+++ crates/mvm-hostd/src/bin/mvm-libkrun-supervisor.rs
@@ -437,29 +437,15 @@ fn run_legacy(cfg: &SupervisorConfig) -> Result<std::convert::Infallible> {
 /// which is the honest answer for a machine that never started.
 static USAGE_STATE_DIR: std::sync::OnceLock<std::path::PathBuf> = std::sync::OnceLock::new();
 
-/// Persist what this supervisor consumed on behalf of its machine.
-///
-/// The VMM shares this process, so the reading covers guest execution together
-/// with device emulation and vsock pumping. Best-effort, like the exit capture
-/// beside it: a workload that already ran must not fail its teardown over
-/// evidence, and a failed probe yields an unavailable metric rather than a zero.
-///
-/// `host_state_bytes` and `wall_ms` are deliberately left unobserved — the host
-/// measures both when it reports the exit, and it is better placed to.
-fn record_self_usage(vm_state_dir: &std::path::Path) {
-    let usage = mvm_core::usage_capture::UsageCapture {
-        cpu_ms: mvm_vmm::host::process_usage::process_cpu_ms_self(),
-        peak_rss_mib: mvm_vmm::host::process_usage::peak_rss_mib_self(),
-        ..mvm_core::usage_capture::UsageCapture::default()
-    };
-    let _ = mvm_core::usage_capture::write_captured(vm_state_dir, &usage);
-}
-
 /// The `atexit` trampoline. Takes the reading against the directory the run loop
 /// was entered for; does nothing if no run loop was ever entered.
+///
+/// The measurement itself lives in `mvm_hostd::supervisor::self_usage`, which is
+/// feature-independent: this binary compiles only under `libkrun-sys`, and a
+/// measurement tested only here would be tested in no lane at all.
 extern "C" fn record_self_usage_at_exit() {
     if let Some(dir) = USAGE_STATE_DIR.get() {
-        record_self_usage(dir);
+        mvm_hostd::supervisor::self_usage::record_self_usage(dir);
     }
 }
 
@@ -505,10 +491,8 @@ fn append_supervisor_breadcrumb(vm_state_dir: &std::path::Path, stage: &str, det
 mod tests {
     use super::{
         append_supervisor_breadcrumb, apply_egress_relay_override, hold_exclusive_image_lock,
-        record_self_usage,
     };
     use libkrun_sys::{BridgeRestartPolicy, KrunContext, NetworkingMode, SupervisorConfig};
-    use mvm_core::usage_capture::{Mechanism, Metric, UsageSource};
 
     fn sample_cfg(networking: NetworkingMode, tenant_id: Option<&str>) -> SupervisorConfig {
         let mut krun = KrunContext::new("vm-1", "/k", "/r");
@@ -595,37 +579,6 @@ mod tests {
         let body = std::fs::read_to_string(dir.path().join("supervisor.lifecycle.log"))
             .expect("read lifecycle log");
         assert!(body.contains("dispatch_config: detail"), "body: {body}");
-    }
-
-    #[test]
-    fn the_supervisor_records_its_own_consumption_as_the_machines() {
-        // The VMM is in-process, so this process's CPU is the machine's CPU plus
-        // this process's own overhead — which is why the mechanism says so.
-        let dir = tempfile::tempdir().expect("tempdir");
-        record_self_usage(dir.path());
-        let captured = mvm_core::usage_capture::read_captured(dir.path());
-        assert_eq!(captured.cpu_ms.source(), UsageSource::Measured);
-        assert!(matches!(
-            captured.cpu_ms,
-            Metric::Measured {
-                mechanism: Mechanism::HostProcessCpu,
-                ..
-            }
-        ));
-        assert_eq!(captured.peak_rss_mib.source(), UsageSource::Measured);
-    }
-
-    #[test]
-    fn the_supervisor_claims_only_the_two_dimensions_it_observes() {
-        // The host fills the state-dir size and the wall span when it reports the
-        // exit. Writing a number for either here would be overwritten at best,
-        // and at worst would contradict the host's own reading.
-        let dir = tempfile::tempdir().expect("tempdir");
-        record_self_usage(dir.path());
-        let captured = mvm_core::usage_capture::read_captured(dir.path());
-        assert_eq!(captured.host_state_bytes, Metric::Unavailable);
-        assert_eq!(captured.wall_ms, Metric::Unavailable);
-        assert_eq!(captured.guest_peak_rss_kib, Metric::Unavailable);
     }
 
     #[test]
